% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{calcf2}
\alias{calcf2}
\title{Compute f2 stats genome-wide}
\usage{
calcf2(f2_blocks)
}
\arguments{
\item{f2_blocks}{A 3d array of blocked f2 statistics}
}
\value{
Returns NxN matrix storing f2 statistics between pairs of populations
}
\description{
This function aggregates f2_blocks to compute a point estimate of f2 genome-wide
}
