% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_from_Relate}
\alias{D_from_Relate}
\title{Function to calculate D statistics from Relate trees.}
\usage{
D_from_Relate(
  file_anc,
  file_mut,
  poplabels,
  pop1,
  pop2,
  pop3,
  pop4,
  file_map = NULL,
  chrs = NULL,
  blgsize = NULL,
  mu = NULL,
  tmin = NULL,
  t = NULL,
  transitions = NULL,
  use_muts = NULL,
  minMAF = NULL
)
}
\arguments{
\item{file_anc}{Filename of anc file. If chrs is specified, this should only be the prefix, resulting in filenames of $\{file_anc\}_chr$\{chr\}.anc(.gz).}

\item{file_mut}{Filename of mut file. If chrs is specified, this should only be the prefix, resulting in filenames of $\{file_anc\}_chr$\{chr\}.anc(.gz).}

\item{poplabels}{Filename of poplabels file}

\item{pop1}{Population 1. Either a population in poplabels or Root.}

\item{pop2}{Population 2. Either a population in poplabels or Root.}

\item{pop3}{Population 3. Either a population in poplabels or Root.}

\item{pop4}{Population 4. Either a population in poplabels or Root.}

\item{file_map}{(Optional) File prefix of recombination map. Not needed if blgsize is given in base-pairs, i.e. blgsize > 100}

\item{chrs}{(Optional) Vector of chromosome IDs}

\item{blgsize}{(Optional) SNP block size in Morgan. Default is 0.05 (5 cM). If blgsize is 100 or greater, if will be interpreted as base pair distance rather than centimorgan distance. If blgsize is negative, every tree is its own block.}

\item{mu}{(Optional) Per base per generation mutation rate to scale f2 values. Default: 1.25e-8}

\item{tmin}{(Optional) Minimum time cutof in generations. Any lineages younger than tmin will be excluded from the analysis. Default: t = 0.}

\item{t}{(Optional) Time cutoff in generations. Default: Inf}

\item{transitions}{(Optional) Set this to FALSE to exclude transition SNPs. Only meaningful with use_muts}

\item{use_muts}{(Optional) Calculate traditional f2 statistics by only using mutations mapped to Relate trees. Default: false.}

\item{minMAF}{(Optional) Minimum frequency cutoff. Default: 1 (i.e. excl singletons)}
}
\value{
Data frame accepted by jackknife function.
}
\description{
This function will calculate D statistics in blocks of prespecified size for all pairs of populations specified in the poplabels file.
Please refer to the Relate documentation for input file formats (https://myersgroup.github.io/relate/).
}
\examples{
file_anc  <- system.file("sim/msprime_ad0.8_split250_1_chr1.anc.gz", package = "twigstats")
file_mut  <- system.file("sim/msprime_ad0.8_split250_1_chr1.mut.gz", package = "twigstats")
poplabels <- system.file("sim/msprime_ad0.8_split250_1.poplabels", package = "twigstats")
file_map  <- system.file("sim/genetic_map_combined_b37_chr1.txt.gz", package = "twigstats")

#Calculate D
D_blocks <- D_from_Relate(file_anc, file_mut, poplabels, "P4", "P3", "P1", "P2", file_map)
print(jackknife(D_blocks))
D_blocks <- D_from_Relate(file_anc, file_mut, poplabels, "P4", "P3", "P1", "PX", file_map)
print(jackknife(D_blocks))

#Calculate D with time cutoff
D_blocks <- D_from_Relate(file_anc, file_mut, poplabels, "P4", "P3", "P1", "PX", file_map, t = 1000)
print(jackknife(D_blocks))

#Calculate D with mutations only (and time cutoff)
D_blocks <- D_from_Relate(file_anc, file_mut, poplabels, "P4", "P3", "P1", "PX", file_map, use_muts = T, t = 1000)
print(jackknife(D_blocks))
}
