% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{f4_ratio}
\alias{f4_ratio}
\alias{jackknife}
\title{Function implementing the F4-ratio statistic}
\usage{
f4_ratio(f2_blocks, popO, popI, pop1, pop2, popX)

jackknife(df_jack)
}
\arguments{
\item{f2_blocks}{A 3d array of blocked f2 statistics}

\item{popO}{Name of outgroup population}

\item{popI}{Name of ingroup population}

\item{pop1}{Name of source that clusters with ingroup}

\item{pop2}{Name of other source}

\item{popX}{Name of target group.}

\item{data}{frame. Three columns called blockID, hj, Dj, storing block ID, weight of block, and statistic without that block}
}
\value{
Returns a data frame with admixture proportion estimates and jacknifed standard errors.

Returns a data table with columns val and se.
}
\description{
This function computes the admixture proportion given five populations. A population history following ((PI,P1),P2,PO) is assumed,
and the target is assumed to be a mixture of proximal sources P1 and P2, i.e. PX = alpha*P2 + (1-alpha)*P1

This function implements a jackknife on the input data.
}
\examples{
#These lines assign file names to variables file_anc, file_mut, poplabels, file_map.
#see https://myersgroup.github.io/relate/getting_started.html#Output for file formats
file_anc  <- system.file("sim/msprime_ad0.8_split250_1_chr1.anc.gz", package = "twigstats")
file_mut  <- system.file("sim/msprime_ad0.8_split250_1_chr1.mut.gz", package = "twigstats")
#see https://myersgroup.github.io/relate/input_data.html for file formats
poplabels <- system.file("sim/msprime_ad0.8_split250_1.poplabels", package = "twigstats")
file_map  <- system.file("sim/genetic_map_combined_b37_chr1.txt", package = "twigstats") #recombination map (three column format)

#Calculate regular f2s between all pairs of populations
f2_blocks1 <- f2_blocks_from_Relate(file_anc, file_mut, poplabels, file_map)
f4_ratio(f2_blocks1, popX="PX", popI="P1", pop1="P2", pop2="P3", popO="P4")

#Use a twigstats cutoff of 500 generations
f2_blocks2 <- f2_blocks_from_Relate(file_anc, file_mut, poplabels, file_map, t = 500)
f4_ratio(f2_blocks2, popX="PX", popI="P1", pop1="P2", pop2="P3", popO="P4") 
}
