% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painting.R
\name{PaintingNNLS}
\alias{PaintingNNLS}
\title{Function to run an NNLS on genome-wide copying proportions to estimate admixture proportions.}
\usage{
PaintingNNLS(df_sum, target_pops, source_pops = NULL)
}
\arguments{
\item{df_sum}{Output of PaintingProfile}

\item{target_pops}{Vector of populations to be fitted}

\item{source_pops}{(Optional) Vector of putative source populations. If not provided, all remaining populations are used.}
}
\value{
Returns a data frame with admixture proportions.
}
\description{
This function computes admixture proprtions using a non-negative least squares on painting profiles obtained from the function PaintingProfile
}
\examples{
library(dplyr)

#This path stores files precomputed using Painting().
path      <- paste0(system.file("test/", package = "twigstats"),"/")
prefix    <- "test" #prefix of files under path

#compute the painting profiles with 100 bootstrap samples and a blocksize of 5cM (5000*0.001)
df <- PaintingProfile(c(paste0(path,prefix,"_painting.txt.gz")), paste0(path,prefix,"_idfile.txt.gz"), nboot = 100, blocksize = 5000)

#compute the NNLS to get admixture proportions
df <- PaintingNNLS(df, target_pops = c("PX"), source_pops = c("P2","P3"))

#Now you can summarize the bootstrap samples 
df \%>\% group_by(target, POP) \%>\% summarize(mean_ancestry = mean(ancestry), sd_ancestry = sd(ancestry)) -> df
print(df)
}
