% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painting.R
\name{PaintingProfile}
\alias{PaintingProfile}
\title{Function to compute genome-wide copying proportions.}
\usage{
PaintingProfile(
  filename_painting,
  filename_idfile,
  nboot,
  blocksize,
  use_IDs = FALSE
)
}
\arguments{
\item{filename_painting}{Vector containing filenames of painting profiles. Output of Painting.}

\item{filename_idfile}{Filename of idfile. Output of Painting.}

\item{nboot}{Number of bootstrap samples.}

\item{blocksize}{Number of blocks to combine for the bootstrap. E.g. if Painting was run with a blgsize of 1e-5 Morgans, blocksize should be 5000 to achieve a blocksize of 5cM.}

\item{use_IDs}{If TRUE, compute profile for each sample. If FALSE (default), compute profile for each group as specified in the second column of the idfile.}
}
\value{
Returns a data frame with copying proportions per bootstrap sample.
}
\description{
This function takes the output of the function Painting and computes the genome-wide 'copying vectors',
i.e. the proportion of the genome copied from each other reference population..
}
\examples{
#This path stores files precomputed using Painting().
path      <- paste0(system.file("test/", package = "twigstats"),"/")
prefix    <- "test" #prefix of files under path

#compute the painting profiles with 10 bootstrap samples and a blocksize of 5cM (5000*0.001)
df <- PaintingProfile(c(paste0(path,prefix,"_painting.txt.gz")), paste0(path,prefix,"_idfile.txt.gz"), nboot = 10, blocksize = 5000)
head(df)
}
