% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{theoretical_zscore}
\alias{theoretical_zscore}
\title{Function to compute a theoretical z-score given source split times and admixture proportions.}
\usage{
theoretical_zscore(t, a, s, alpha)
}
\arguments{
\item{t}{Twigstats cutoff time in units of 2Ne generations}

\item{a}{Admixture time in units of 2Ne generations}

\item{s}{Source split time in units of 2Ne generations}

\item{alpha}{Admixture proportions}
}
\value{
Returns the theoretical z-score value.
}
\description{
This function computes the theoretical z-score of an f4-statistic of the form f4(PO,P2,PX,P1) 
as a function of the admixture time a, source (P1 & P2) split time s, and admixture proportion alpha. 
It assumes a single lineage is sampled from each population.
}
