% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TMRCA_from_Relate}
\alias{TMRCA_from_Relate}
\title{Function to calculates mean TMRCAs from Relate trees for pairs of populations specified in poplabels.}
\usage{
TMRCA_from_Relate(
  file_anc,
  file_mut,
  poplabels,
  file_out,
  file_map = NULL,
  chrs = NULL,
  t = NULL,
  blgsize = NULL
)
}
\arguments{
\item{file_anc}{Filename of anc file. If chrs is specified, this should only be the prefix, resulting in filenames of $\{file_anc\}_chr$\{chr\}.anc(.gz).}

\item{file_mut}{Filename of mut file. If chrs is specified, this should only be the prefix, resulting in filenames of $\{file_anc\}_chr$\{chr\}.anc(.gz).}

\item{poplabels}{Filename of poplabels file}

\item{file_out}{Filename of output file.}

\item{file_map}{(Optional) File prefix of recombination map. Not needed if blgsize is given in base-pairs, i.e. blgsize > 100}

\item{chrs}{(Optional) Vector of chromosome IDs}

\item{t}{(Optional) Time cutoff in generations. Any coalescences older that t will be set to t in the analysis. Default: t = Inf.}

\item{blgsize}{(Optional) SNP block size in Morgan. Default is 0.05 (50 cM). If blgsize is 100 or greater, if will be interpreted as base pair distance rather than centimorgan distance.}
}
\value{
3d array of dimension #groups x #groups x #blocks. Analogous to output of f2_from_geno in admixtools.
}
\description{
This function will calculate TMRCAs in blocks of prespecified size for all pairs of populations specified in the poplabels file.
Please refer to the Relate documentation for input file formats (https://myersgroup.github.io/relate/).
}
\examples{
file_anc  <- system.file("sim/msprime_ad0.8_split250_1_chr1.anc.gz", package = "twigstats")
file_mut  <- system.file("sim/msprime_ad0.8_split250_1_chr1.mut.gz", package = "twigstats")
poplabels <- system.file("sim/msprime_ad0.8_split250_1.poplabels", package = "twigstats")
file_map  <- system.file("sim/genetic_map_combined_b37_chr1.txt", package = "twigstats")

#Calculate f2s between all pairs of populations
TMRCA_from_Relate(file_anc, file_mut, poplabels, file_out = "test", file_map)
}
\keyword{internal}
