% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PaintingProfile}
\alias{PaintingProfile}
\title{Function to compute genome-wide copying proportions.}
\usage{
PaintingProfile(
  filename_painting,
  filename_idfile,
  nboot,
  blocksize,
  use_IDs = FALSE
)
}
\arguments{
\item{filename_painting}{A vector containing filenames of painting profiles. These are the outputs of Painting.}

\item{filename_idfile}{The filename of the idfile, which is an output of Painting.}

\item{nboot}{Number of bootstrap samples to generate.}

\item{blocksize}{Number of blocks to combine for the bootstrap. For example, if Painting was run with a blgsize of 1e-5 Morgans, blocksize should be 5000 to achieve a blocksize of 5cM.}

\item{use_IDs}{If TRUE, compute profiles for each sample. If FALSE (default), compute profiles for each group as specified in the second column of the idfile.}
}
\value{
A DataFrame with the copying proportions per bootstrap sample.
}
\description{
This function takes the output of the function Painting and computes the genome-wide 'copying vectors',
i.e., the proportion of the genome copied from each reference population.
}
\examples{
# Example usage in R:
path      <- paste0(system.file("test/", package = "twigstats"),"/")
prefix    <- "test" # prefix of files under path
df <- PaintingProfile(c(paste0(path, prefix, "_painting.txt.gz")),
                         paste0(path, prefix, "_idfile.txt.gz"),
                         nboot = 10, blocksize = 5000)
head(df)
}
